# apps/web/public/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "SlyncPay API",
    "version": "2.0.0",
...
  },
  "servers": [
    { "url": "https://slyncpay-api.onrender.com", "description": "Production" }
  ],
  "security": [{ "bearerAuth": [] }],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "API Key"
      }
    },
    "schemas": {
      "Error": {
        "type": "object",
        "properties": {
          "error": { "type": "string", "example": "validation_error" },
          "message": { "type": "string", "example": "Field xyz is required" },
          "statusCode": { "type": "integer", "example": 400 }
        }
      },
...
      "Entity": {
        "type": "object",
        "properties": {
          "id": { "type": "string", "format": "uuid" },
          "name": { "type": "string" },
          "einLast4": { "type": ["string", "null"] },
          "state": { "type": ["string", "null"] },
          "status": { "type": "string", "enum": ["pending", "active", "suspended"] },
          "taxType": { "type": "string", "enum": ["1099", "w2"], "description": "Determines whether this entity runs 1099 contractor payments or W-2 employee payroll. Locked at create time." },
          "createdAt": { "type": "string", "format": "date-time" },
          "updatedAt": { "type": "string", "format": "date-time" }
        }
      },
      "EntityDetail": {
        "allOf": [
          { "$ref": "#/components/schemas/Entity" },
          {
            "type": "object",
            "properties": {
              "wingspan": {
                "type": "object",
                "description": "Identifiers that map this entity to the underlying payment processor (Wingspan). Useful for debugging and for tenants who hold a Wingspan parent-org account.",
                "properties": {
                  "environment": { "type": "string", "enum": ["live", "test"] },
                  "organizationUserId": { "type": "string" },
                  "childUserId": { "type": ["string", "null"] },
                  "childUserEmail": { "type": ["string", "null"] },
                  "uiBaseUrl": { "type": "string", "description": "Wingspan UI host for the current environment." }
                }
              }
            }
          }
        ]
      },
      "ProvisioningStatus": {
        "type": "object",
        "description": "Tracks async provisioning of an entity (Wingspan child-user setup runs in the background after POST /v1/entities).",
        "properties": {
          "status": { "type": "string", "enum": ["not_started", "queued", "running", "completed", "failed"] },
          "currentStep": { "type": ["string", "null"] },
          "stepsCompleted": { "anyOf": [{ "type": "array", "items": { "type": "string" } }, { "type": "null" }] },
          "lastError": { "type": ["string", "null"] },
          "updatedAt": { "type": "string", "format": "date-time" }
        }
      },
...
    }
  },
...
  "paths": {
...
    "/v1/entities": {
      "get": {
        "tags": ["Entities"],
        "summary": "List your entities",
        "responses": {
          "200": {
            "description": "OK",
            "content": { "application/json": { "schema": { "type": "array", "items": { "$ref": "#/components/schemas/Entity" } } } }
          }
        }
      },
      "post": {
        "tags": ["Entities"],
        "summary": "Create an entity",
        "description": "Creates a new legal entity (one per EIN). Provisioning runs asynchronously — poll `GET /v1/entities/{id}/provisioning-status` until the entity's `status` flips to `active` before paying workers against it.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["name", "ein"],
                "properties": {
                  "name": { "type": "string", "minLength": 1, "maxLength": 100 },
                  "ein": { "type": "string", "pattern": "^\\d{2}-\\d{7}$", "description": "Format XX-XXXXXXX." },
                  "state": { "type": "string", "minLength": 2, "maxLength": 2 },
                  "taxType": { "type": "string", "enum": ["1099", "w2"], "default": "1099", "description": "Locks the entity's classification. W-2 entities accept worker engagements but payable creation returns 501 until the W-2 runtime ships." }
                }
              },
              "example": { "name": "NurseIO, Inc.", "ein": "83-3439766", "state": "AZ", "taxType": "1099" }
            }
          }
        },
        "responses": {
          "201": { "description": "Created (provisioning runs in background)", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Entity" } } } },
          "402": { "description": "Plan entity limit reached", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } }
        }
      }
    },
    "/v1/entities/{id}": {
      "get": {
        "tags": ["Entities"],
        "summary": "Get an entity",
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } }],
        "responses": {
          "200": { "description": "OK", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/EntityDetail" } } } },
          "404": { "description": "Not found", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } }
        }
      },
      "delete": {
        "tags": ["Entities"],
        "summary": "Delete an entity",
        "description": "Removes an entity that has no workers or payables attached.",
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } }],
        "responses": {
          "204": { "description": "Deleted" },
          "409": { "description": "Entity has dependent records", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } }
        }
      }
    },
    "/v1/entities/{id}/provisioning-status": {
      "get": {
        "tags": ["Entities"],
        "summary": "Check entity provisioning status",
        "description": "Polled after `POST /v1/entities` to know when the entity's processor account is ready. Once the entity's main `status` reads `active`, you can stop polling and start creating payables against it.",
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } }],
        "responses": {
          "200": { "description": "OK", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ProvisioningStatus" } } } }
        }
      }
    },
...
  }
}